- combined two instrumentations into a single java class
